Renamed traits to match Symfony coding standards and renamed integration methods to integrationConfiguration
